requirements.txt
# Medical Image Classifier — Dependencies
# Install PyTorch with CUDA 11.8 first:
#   pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
# Then install the rest:
#   pip install -r requirements.txt

grad-cam
scikit-learn
matplotlib
seaborn
streamlit
pillow
pandas
tqdm
